+++ calico/felix/test/test_felix.py
@@ -72,6 +72,7 @@ class TestBasic(BaseTestCase):
         m_MasqueradeManager.return_value.greenlet = mock.Mock()
         m_UpdateSplitter.return_value.greenlet = mock.Mock()
         m_config = mock.Mock(spec=config.Config)
+        m_config.ETCD_ADDR = "localhost:4001"
         m_config.HOSTNAME = "myhost"
         m_config.IFACE_PREFIX = "tap"
         m_config.METADATA_IP = "10.0.0.1"